% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_section.R
\name{duplicate_section}
\alias{duplicate_section}
\title{Duplicate section}
\usage{
duplicate_section(section, name, prefix = NULL, suffix = NULL)
}
\arguments{
\item{section}{section element according formatted according to the json schema in Questions.json}

\item{name}{name of the new section}

\item{prefix}{character to add to the beginning of existing ids to make them unique on dynamic rendering}

\item{suffix}{character to add to the end of existing ids to make them unique on dynamic rendering}
}
\description{
The function duplicates a section and assigns unique ids to dynamically to the dynamically created elements.
}
